+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+#### Changed
+
+- Changed `for`'s callback to return `ByteStream m x`, to clarify that
+  it is not used.
+
 ## 0.2.3 (2022-08-18)
 
 #### Added